fix: 64-bit offsets for TREC files >4GB
Two related 32-bit integer overflows would cause Zettair to record
wrong byte offsets for documents past 4GB into the source TREC file:

1. struct docmap_entry.offset was unsigned int, so any offset > 4GB
   wrapped on assignment from off_t. Reading back gave the wrapped
   (32-bit) value, and the summariser would seek into the wrong part
   of the TREC, returning text from a different document.

2. In index.c::index_add, bytes_read/last_pos/curr_pos were unsigned
   int. Once the input stream crossed 4GB, these wrapped, and even
   though docmap_add() takes off_t for offset, the caller was passing
   the truncated value.

Reproduced locally with a 4.8GB synthetic TREC: doc 99999 (at byte
offset 5,131,622,786) was returning the document found at byte
offset 836,655,490 (= 5131622786 mod 2^32). After the fix all docs
return correct content.

This is what was causing the production bug where Wikipedia query
summaries showed text from unrelated articles after the 1.5M corpus
rebuild produced a 13GB TREC.

+++ devel/src/include/_docmap.h
@@ -31,7 +31,7 @@ struct fdset;
 struct docmap_entry {
     unsigned int docno;               /* doc number */
     unsigned int fileno;              /* repository file number */
-    unsigned int offset;              /* offset within repository file */
+    off_t offset;                     /* offset within repository file (64-bit, supports >4GB TREC) */
     unsigned int dwords;              /* number of distinct words in doc */
     unsigned int words;               /* number of words in doc */
     double weight;                     /* the cosine weight of the doc */
